Reduce complexity by using :0 instead of vnc
By setting DISPLAY=:0 we don't need to start a vncserver. We render
directly on the local X Server making TurboVNC and VirtualGL obsolete.
Therefore we no longer need the F* Init-Scripts. Setting up the graphic
driver is still mandatory however.
The disadvantage is we can no longer remotly view the output - a feature
which we didn't intend to use anyways. The advantage is a reduction of
complexity. Include glxspeheres64 from the VirtualGL package for a simple
3D-benchmarking test.